add a visual servoing to try with the camera
This function performs template matching as a convolution to check the position of the desired target image recorded during the demonstration (cropped in the center) in the current complete image received from the camera sensor. 
The robot needs to move in x and y in the camera frame proportionally to the estimated distance, no minus required. We must convert that x and y back in the base frame before adding it to the current robot action.